test: add test for `ecaviar_colocalisation`
